VYBES-111: implement usequery
