+++ app/_layout.tsx
@@ -1,4 +1,5 @@
 import {DarkTheme, DefaultTheme, ThemeProvider} from '@react-navigation/native';
+import {QueryClient, QueryClientProvider} from '@tanstack/react-query';
 import {useFonts} from 'expo-font';
 import {Stack} from 'expo-router';
 import * as SplashScreen from 'expo-splash-screen';
@@ -13,6 +14,7 @@ import SpaceMono from '../assets/fonts/SpaceMono-Regular.ttf';
 SplashScreen.preventAutoHideAsync();
 
 export default function RootLayout() {
+  const queryClient = new QueryClient();
   const colorScheme = useColorScheme();
   const [loaded] = useFonts({
     SpaceMono,
@@ -29,16 +31,18 @@ export default function RootLayout() {
   }
 
   return (
-    <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
-      <Stack>
-        <Stack.Screen name='index' options={{headerShown: false}} />
-        <Stack.Screen name='login' options={{headerShown: false}} />
-        <Stack.Screen name='debug' options={{headerShown: false}} />
-        <Stack.Screen name='(tabs)' options={{headerShown: false}} />
-        <Stack.Screen name='playlist' options={{headerShown: false}} />
-        <Stack.Screen name='+not-found' />
-      </Stack>
-      <StatusBar style='auto' />
-    </ThemeProvider>
+    <QueryClientProvider client={queryClient}>
+      <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
+        <Stack>
+          <Stack.Screen name='index' options={{headerShown: false}} />
+          <Stack.Screen name='login' options={{headerShown: false}} />
+          <Stack.Screen name='debug' options={{headerShown: false}} />
+          <Stack.Screen name='(tabs)' options={{headerShown: false}} />
+          <Stack.Screen name='playlist' options={{headerShown: false}} />
+          <Stack.Screen name='+not-found' />
+        </Stack>
+        <StatusBar style='auto' />
+      </ThemeProvider>
+    </QueryClientProvider>
   );
 }
