+++ app/debug.tsx
@@ -1,67 +1,27 @@
 /* eslint-disable no-console */
 import Screen from '@/components/Screen';
-import {useEffect, useState} from 'react';
+import {useGetTracksQuery} from '@/hooks/network';
 import {SafeAreaView, StyleSheet, Text, View} from 'react-native';
-import {Network} from '../hooks/network/Network';
-import {Track} from '../types';
-
-const getTracks = async () => {
-  try {
-    // Using the Network class to make a GET request to the Supabase REST API
-    const response = await Network.get<Track[]>('/rest/v1/tracks');
-    return response;
-  } catch (error) {
-    console.error('Error fetching tracks:', error);
-    throw error;
-  }
-};
 
 export default function DebugScreen() {
-  const [data, setData] = useState<Track[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
+  const {data, isLoading, error} = useGetTracksQuery();
+  console.log('tracksData', {
+    data,
+    isLoading,
+    error,
+  });
 
-  useEffect(() => {
-    const fetchTracks = async () => {
-      try {
-        setLoading(true);
-        setError(null);
-
-        const result = await getTracks();
-
-        if (Array.isArray(result)) {
-          const tracks: Track[] = [];
-          result.forEach((element) => {
-            tracks.push(element ?? {track_name: 'missing'});
-          });
-          setData(tracks);
-        } else {
-          setError('Invalid response format');
-        }
-      } catch (err) {
-        setError(err instanceof Error ? err.message : 'Unknown error occurred');
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchTracks();
-  }, []);
-
-  // parse data
   console.log('hi', data);
 
   return (
     <SafeAreaView>
       <Screen title='Debug' subtitle='Debug' iconName='back'>
         <View style={styles.responseContainer}>
-          {loading && <Text style={styles.response}>Loading...</Text>}
-          {error && (
-            <Text style={[styles.response, styles.error]}>{error}</Text>
-          )}
-          {!loading && !error && (
+          {isLoading && <Text style={styles.response}>Loading...</Text>}
+          {error && <Text style={[styles.response, styles.error]}>error</Text>}
+          {!isLoading && !error && (
             <Text style={styles.response}>
-              {data.length > 0
+              {data && data.length > 0
                 ? `Loaded ${data.length} tracks`
                 : 'No tracks found'}
             </Text>
