+++ package.json
@@ -20,6 +20,7 @@
     "@react-navigation/bottom-tabs": "^7.2.0",
     "@react-navigation/native": "^7.0.14",
     "@supabase/supabase-js": "^2.53.0",
+    "@tanstack/react-query": "^5.84.0",
     "@types/bun": "^1.2.19",
     "expo": "~52.0.37",
     "expo-blur": "~14.0.3",
@@ -33,6 +34,7 @@
     "expo-symbols": "~0.2.2",
     "expo-system-ui": "~4.0.8",
     "expo-web-browser": "~14.0.2",
+    "g": "^2.0.1",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-native": "0.76.7",
